ai_prd_agent_full/requirements.txt
fastapi
uvicorn
jinja2
pydantic
requests
markdown2
weasyprint
python-dotenv